fix: stock con unidades de medida, modal cancelación y MercadoPago
- Cambiar stock_cantidad de int a Decimal(10,3) en ingredientes
- Agregar conversión de unidades (g↔kg, mL↔L, u↔doc) en checkout
- Reemplazar prompt() nativo por modal con textarea en ambos frontends
- Corregir SDK MercadoPago: usar RequestOptions para idempotency_key
- Quitar auto_return (requiere HTTPS) para compatibilidad con localhost
- Reutilizar pago pendiente existente en vez de duplicar external_reference
- Filtrar productos sin stock del listado público (stock_cantidad > 0)
- Usar z.coerce.number() en schema de producto para Decimals del backend
- Permitir cantidades decimales en receta de ingredientes

+++ backend/app/modules/ingredientes/schema.py
@@ -1,4 +1,5 @@
 from typing import Optional
+from decimal import Decimal
 from datetime import datetime
 from app.modules.ingredientes.model import IngredienteBase
 
@@ -14,5 +15,5 @@ class IngredienteUpdate(IngredienteBase):
     nombre: Optional[str] = None
     descripcion: Optional[str] = None
     es_alergeno: Optional[bool] = None
-    stock_cantidad: Optional[int] = None
+    stock_cantidad: Optional[Decimal] = None
     unidad_medida_id: Optional[int] = None
